womir: widen opcode register indices to u32 in executors

+++ extensions/womir_circuit/src/base_alu/execution.rs
@@ -75,9 +75,9 @@ pub(super) struct BaseAluPreCompute {
     /// Second operand value (if immediate) or register index (if register)
     c: u32,
     /// Result register index
-    a: u8,
+    a: u32,
     /// First operand register index
-    b: u8,
+    b: u32,
 }
 
 impl<A, const NUM_LIMBS: usize, const LIMB_BITS: usize> BaseAluExecutor<A, NUM_LIMBS, LIMB_BITS> {
@@ -104,8 +104,8 @@ impl<A, const NUM_LIMBS: usize, const LIMB_BITS: usize> BaseAluExecutor<A, NUM_L
             } else {
                 c_u32
             },
-            a: a.as_canonical_u32() as u8,
-            b: b.as_canonical_u32() as u8,
+            a: a.as_canonical_u32(),
+            b: b.as_canonical_u32(),
         };
         Ok(is_imm)
     }
@@ -226,7 +226,7 @@ unsafe fn execute_e12_impl<
     const { assert!(NUM_LIMBS == 4 || NUM_LIMBS == 8) };
 
     let fp = exec_state.memory.fp::<F>();
-    let rs1 = exec_state.vm_read::<u8, NUM_LIMBS>(RV32_REGISTER_AS, fp + (pre_compute.b as u32));
+    let rs1 = exec_state.vm_read::<u8, NUM_LIMBS>(RV32_REGISTER_AS, fp + pre_compute.b);
     let rs2 = if IS_IMM {
         sign_extend_u32::<NUM_LIMBS>(pre_compute.c)
     } else {
@@ -240,7 +240,7 @@ unsafe fn execute_e12_impl<
     ));
     let result = OP::compute(a, b).to_le_bytes();
     let rd: [u8; NUM_LIMBS] = std::array::from_fn(|i| result[i]);
-    exec_state.vm_write::<u8, NUM_LIMBS>(RV32_REGISTER_AS, fp + (pre_compute.a as u32), &rd);
+    exec_state.vm_write::<u8, NUM_LIMBS>(RV32_REGISTER_AS, fp + pre_compute.a, &rd);
     let pc = exec_state.pc();
     exec_state.set_pc(pc.wrapping_add(DEFAULT_PC_STEP));
 }
